year-view: Mark working days according to Locale
Currently, on the year view, Saturday and Sunday are marked
as non working days by default, despite the Locale settings and the
fact that in some Countries Sunday and/or Saturday are workdays.

This patch introduces a new function "is_workday" to check
whether a day is a working day according to the Locale data,
specifically the "territory" part of it, and a Hash table of
Territories whose non working days are other than Saturday and Sunday.